Implemented holding module and holding dto with manual mapping using builder .

+++ project/src/main/java/com/portfolioforecasting/dto/request/HoldingRequest.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -17,7 +18,7 @@ public class HoldingRequest {
     private String stockSymbol;
     private String companyName;
     private Integer quantity;
-    private Double buyPrice;
+    private BigDecimal buyPrice;
     private LocalDate purchaseDate;
     private Long portfolioId;
 }
